#5-refact: 필요없는 파일 삭제(일반적인 웹소켓에서는 필요하지만 STOMP에서는 필요 없음), messenger 관련 컨트롤러의 이름을 각각의 역할에 맞게 직관적으로 이름 수정
