+++ src/main/java/com/example/demo/controller/messenger/MessageController.java
@@ -1,93 +1,45 @@
 package com.example.demo.controller.messenger;
 
-import com.example.demo.dto.messenger.ChatRoomRequestDTO;
-import com.example.demo.dto.messenger.ChatRoomResponseDTO;
+import com.example.demo.entity.messenger.Message;
 import com.example.demo.service.messenger.ChatRoomService;
+import com.example.demo.service.messenger.ChatMessageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.simp.SimpMessageSendingOperations;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
 
 @RestController()
-@RequestMapping("/messeneger")
 @RequiredArgsConstructor
 @Slf4j
 public class MessageController {
 
     private final ChatRoomService chatRoomService;
-    private static final Logger logger = LoggerFactory.getLogger(MessageController.class);
+    private final ChatMessageService chatMessageService;
 
-    // 채팅방 생성
-    @PostMapping("/chatRoom")
-    public ResponseEntity<ChatRoomResponseDTO> createChatRoom(@RequestBody ChatRoomRequestDTO.CreateDTO requestDTO) {
-        try {
-            ChatRoomResponseDTO responseDTO = chatRoomService.createChatRoom(requestDTO);
-            return ResponseEntity.ok(responseDTO);
-        } catch (Exception e) {
-            logger.error("Error creating chat room", e);
-            return ResponseEntity.internalServerError().body(null);
-        }
+    private final SimpMessageSendingOperations simpMessageSendingOperations;
+
+    @MessageMapping("/hello") // 1. 클라이언트에서 /pub/hello로 메시지 발행
+    public void message(Message message){
+        // 2. 메시지에 정의된 채널 id에 메시지 보냄.
+        // /sub/channel/채널아이디 에 구독중인 클라이언트에게 메시지를 보냄
+        log.info("Received Message: {}", message);
+        simpMessageSendingOperations.convertAndSend("/sub/channel/" + message.getChannelId(), message);
     }
 
-    // 사용자가 참여한 모든 채팅방 조회
-    // localhost:8080/messeneger/chatRooms?userId=minbory925@gmail.com
-    /*
-    현재는 위의 방식처럼 뒤에 userId를 넣어줘야 하지만 추후 JWT 토큰 인증m 추가하면 삭제될 예정임
-     */
-    @GetMapping("/chatRooms")
-    public ResponseEntity<List<ChatRoomResponseDTO>> getAllChatRooms(@RequestParam("userId") String userId) {
-        try {
-            List<ChatRoomResponseDTO> responseDTOList = chatRoomService.findAllChatRoomsByUserId(userId);
-            return ResponseEntity.ok(responseDTOList);
-        } catch (Exception e) {
-            logger.error("Error retrieving chat rooms", e);
-            return ResponseEntity.internalServerError().body(null);
-        }
-    }
-
-    // 특정 채팅방 조회
-    @GetMapping("/chatRoom/{roomId}")
-    public ResponseEntity<ChatRoomResponseDTO> getChatRoomById(@PathVariable("roomId") Long roomId) {
-        try {
-            ChatRoomResponseDTO responseDTO = chatRoomService.findChatRoomById(roomId);
-            return ResponseEntity.ok(responseDTO);
-        } catch (Exception e) {
-            logger.error("Error retrieving chat room", e);
-            return ResponseEntity.internalServerError().body(null);
-        }
-    }
-
-    // 특정 채팅방에 사용자 초대
-    @PostMapping("/chatRoom/{roomId}")
-    public ResponseEntity<Void> inviteUserToChatRoom(@PathVariable("roomId") Long roomId, @RequestBody ChatRoomRequestDTO.InviteDTO inviteRequest) {
-        try {
-            chatRoomService.inviteUserToChatRoom(roomId, inviteRequest);
-            return ResponseEntity.ok().build();
-        } catch (Exception e) {
-            logger.error("Error inviting user to chat room", e);
-            return ResponseEntity.internalServerError().build();
-        }
-    }
-
-    // 특정 채팅방 나가기 (사용자 삭제)
-     /*
-    모든 채팅방 조회와 마찬가지로
-    DELETE /messages/chatRoom/{room_id}?userId={userId}
-    userId를 입력해야 하지만 추후 JWT 토큰 인증을 넣으면 필요없어짐 (API 명세대로 구현할 수 있음)
-     */
-    @DeleteMapping("/chatRoom/{roomId}")
-    public ResponseEntity<Void> leaveChatRoom(@PathVariable("roomId") Long roomId, @RequestParam("userId") String userId) {
-        try {
-            chatRoomService.leaveChatRoom(roomId, userId);
-            return ResponseEntity.ok().build();
-        } catch (Exception e) {
-            logger.error("Error leaving chat room", e);
-            return ResponseEntity.internalServerError().build();
-        }
-    }
+//    @MessageMapping("/chat.sendMessage")
+//    @SendTo("/topic/public")
+//    public ChatMessage sendMessage(@Payload ChatMessage chatMessage) {
+//        chatMessage.setSendAt(LocalDateTime.now());
+//        chatMessageService.saveMessage(chatMessage);
+//        return chatMessage;
+//    }
+//
+//    @MessageMapping("/chat.addUser")
+//    @SendTo("/topic/public")
+//    public ChatMessage addUser(@Payload ChatMessage chatMessage, SimpMessageHeaderAccessor headerAccessor) {
+//        headerAccessor.getSessionAttributes().put("username", chatMessage.getUser().getUserId());
+//        return chatMessage;
+//    }
 }
